Renamed some files to better match routes
