+++ frontend/src/views/Retron5EraseCart.vue
@@ -1,17 +1,17 @@
 <template>
   <div class="home">
-    <Retron5EraseCartridge/>
+    <Retron5EraseSave/>
   </div>
 </template>
 
 <script>
 // @ is an alias to /src
-import Retron5EraseCartridge from '@/components/Retron5EraseCartridge.vue';
+import Retron5EraseSave from '@/components/Retron5EraseSave.vue';
 
 export default {
   name: 'Retron5EraseCart',
   components: {
-    Retron5EraseCartridge,
+    Retron5EraseSave,
   },
 };
 </script>
